Minispade breakout; first ring page pass in Ember.
