+++ priv/admin/js/templates/application.hbs
@@ -23,7 +23,7 @@
                 <div id="split-bar"></div>
             </td>
 
-            <td id="content-well">{{ outlet }}</td>
+            <td id="content-well">{{outlet}}</td>
         </tr>
     </table>
 </div><!-- #wrapper -->
