+++ priv/admin/js/templates/partition_filter.hbs
@@ -1,0 +1,8 @@
+<div id="ring-filter" class="right">
+    <h4 class="gui-headline">Filter</h4>
+    <div class="gui-dropdown-wrapper">
+        <div class="gui-dropdown-bg gui-text smaller"></div>
+        <div class="gui-dropdown-cap left"></div>
+        {{view RiakControl.PartitionFilterSelectView id="filter" classNames="gui-dropdown" contentBinding="this" optionLabelPath="content.name" optionValuePath="content.name"}}
+    </div>
+</div>
