+++ priv/admin/js/templates/ring.hbs
@@ -1,108 +1,60 @@
-<div id="ring-filter" class="right">
-    <!-- Will potentially implement something like this after we get some feedback on how users would like to use the filter.
-    <div class="gui-checkbox-wrapper">
-        <input id="primary-nodes" class="gui-checkbox" type="checkbox" name="ringfilter" checked="checked" />
-        <span class="gui-text">Primary&nbsp;&nbsp;&nbsp;</span>
-    </div>
-    <div class="gui-checkbox-wrapper">
-        <input id="fallback-nodes" class="gui-checkbox" type="checkbox" name="ringfilter" checked="checked" />
-        <span class="gui-text">Fallback&nbsp;&nbsp;&nbsp;</span>
-    </div>
-    -->
-    <!--
-    <div class="gui-checkbox-wrapper">
-        <input id="handoff-nodes" class="gui-checkbox" type="checkbox" name="ringfilter" checked="checked" />
-        <span class="gui-text">Handoff&nbsp;&nbsp;&nbsp;</span>
-    </div>
-    -->
-    <h4 class="gui-headline">Filter</h4>
-    <div class="gui-dropdown-wrapper">
-        <div class="gui-dropdown-bg gui-text smaller"></div>
-        <div class="gui-dropdown-cap left"></div>
-        <select id="filter" class="gui-dropdown">
-            <option value="/admin/ring/partitions">All</option>
-        </select>
-    </div>
-</div>
+{{outlet partitionFilter}}
 
 <h1 id="ring-headline" class="gui-headline-bold">
     Ring View
     <span id="total-number" class="gui-text"></span>
 </h1>
 
-<ul class="pagination gui-text">
-    <!-- The end result will look like this:
-    <li name="prev"><span class="paginator">Prev</span></li>
-    <li name="1"><span class="paginator pageNumber active">1</span></li>
-    <li name="2"><span class="paginator pageNumber">2</span></li>
-    <li name="3"><span class="paginator pageNumber">3</span></li>
-    <li name="next"><span class="paginator">Next</span></li>
-    -->
-</ul>
+<ul class="pagination gui-text"> </ul>
 <div class="cut"></div>
 
-<div id="partition-list" class="hide has-cut">
+<div id="partition-list" class="has-cut">
     <table class="list-table" id="ring-table">
         <thead>
-        <tr class="table-head has-cut">
-            <td><h3 class="gui-headline">#</h3></td>
-            <td><h3 class="gui-headline">Owner Node</h3></td>
-            <td><h3 class="gui-headline">KV</h3></td>
-            <td><h3 class="gui-headline">Pipe</h3></td>
-            <td><h3 class="gui-headline">Search</h3></td>
-        </tr>
-        <tr id="no-matches" class="hide">
-            <td colspan="5" class="gui-text">No partitions matched your filter.</td>
-        </tr>
-        <tr class="spinner-box">
-            <td colspan="5"><img id="ring-spinner" class="spinner" src="/admin/ui/images/spinner.gif"></td>
-        </tr>
+            <tr class="table-head has-cut">
+                <td><h3 class="gui-headline">#</h3></td>
+                <td><h3 class="gui-headline">Owner Node</h3></td>
+                <td><h3 class="gui-headline">KV</h3></td>
+                <td><h3 class="gui-headline">Pipe</h3></td>
+                <td><h3 class="gui-headline">Search</h3></td>
+            </tr>
         </thead>
-        <tbody id="ring-table-body">
-        
-        </tbody>
+        {{#collection RiakControl.PartitionView contentBinding="this"}}
+          {{#with view.content}}
+          <td class="partition-number gui-text">{{i}}</td>
+          <td class="owner-box gui-text">
+              <div class="gui-field gui-text">
+                  <div class="gui-field-bg">
+                      <div class="owner gui-field-input">{{node}}</div>
+                      <div class="partition-index hide">{{index}}</div>
+                  </div>
+                  <div class="gui-field-cap-left"></div>
+                  <div class="gui-field-cap-right"></div>
+              </div>
+          </td>
+          {{/with}}
+          {{#with view}}
+          <td class="kv-box gui-text">
+              <a {{bindAttr class="kvIndicator lightClasses"}}>
+                  <span class="kv-status">{{kvStatus}}</span>
+                  <span class="hide fallback-to"></span>
+              </a>
+          </td>
+          <td class="pipe-box gui-text">
+              <a {{bindAttr class="pipeIndicator lightClasses"}}>
+                  <span class="pipe-status">{{pipeStatus}}</span>
+                  <span class="hide fallback-to"></span>
+              </a>
+          </td>
+          <td class="search-box gui-text">
+              <a {{bindAttr class="searchIndicator lightClasses"}}>
+                  <span class="search-status">{{searchStatus}}</span>
+                  <span class="hide fallback-to"></span>
+              </a>
+          </td>
+          {{/with}}
+        {{/collection}}
     </table>
 </div>
 
 <ul class="pagination gui-text pagination-bottom"></ul>
-
-
-
-<!-- partition template -->
-<table class="hide">
-    <tr class="partition partition-template">
-        <td class="partition-number gui-text"></td>
-        <td class="owner-box gui-text">
-            <div class="gui-field gui-text">
-                <div class="gui-field-bg">
-                    <div class="owner gui-field-input"></div>
-                    <div class="partition-index hide"></div>
-                </div>
-                <div class="gui-field-cap-left"></div>
-                <div class="gui-field-cap-right"></div>
-            </div>
-        </td>
-        <td class="kv-box gui-text">
-            <a class="gui-light kv-light gray">
-                <span class="kv-status">Disabled</span>
-                <span class="hide fallback-to"></span>
-            </a>
-        </td>
-        <td class="pipe-box gui-text">
-            <a class="gui-light pipe-light gray">
-                <span class="pipe-status">Disabled</span>
-                <span class="hide fallback-to"></span>
-            </a>
-        </td>
-        <td class="search-box gui-text">
-            <a class="gui-light search-light gray">
-                <span class="search-status">Disabled</span>
-                <span class="hide fallback-to"></span>
-            </a>
-        </td>
-    </tr>
-</table>
-<!-- end partition template -->
-
-
-
