+++ rebar.config
@@ -26,5 +26,5 @@
     {out_dir,    "priv/admin/js"},
     {target,     "Ember.TEMPLATES"},
     {compiler,   "Ember.Handlebars.compile"},
-    {templates,  [{"templates", ["application", "snapshot", "cluster", "ring", "mapreduce"]}]}
+    {templates,  [{"templates", ["application", "snapshot", "cluster", "ring", "mapreduce", "partition_filter"]}]}
 ]}.
